+++ CHANGELOG.md
@@ -10,6 +10,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - Add field quotesManagedBy on appSettings to handle splitting quotes
+- Process splitting quote by seller if it accepts to manage quotes
+- Notify seller with quote reference data as payload
 
 ## [2.6.4] - 2024-10-31